+++ src/main/java/proguard/classfile/kotlin/visitor/AllValueParameterVisitor.java
@@ -51,7 +51,7 @@ implements KotlinMetadataVisitor,
     {
         kotlinDeclarationContainerMetadata.functionsAccept(clazz, this);
         kotlinDeclarationContainerMetadata.accept(clazz,
-                                                  new AllKotlinPropertiesVisitor(this));
+                                                  new AllPropertyVisitor(this));
     }
 
     @Override